refactor: migrate ProfileWidget and PRReviewer to CSS modules
