+++ src/components/profile-widget/profile-widget.tsx
@@ -1,8 +1,7 @@
-import { Label } from "../ui/label";
-import { Image } from "../ui/image";
-import Platform from "../helpers/Platform";
-import { useAvatar } from "../hooks/useAvatar";
-import { HighlightenText } from "./HighlightenText";
+import Platform from "../../helpers/Platform";
+import { useAvatar } from "../../hooks/useAvatar";
+import { HighlightenText } from "../HighlightenText";
+import styles from "./profile-widget.module.css";
 
 interface IProps {
     nameFull: string;
@@ -23,10 +22,10 @@ export function ProfileWidget({ avatarUrl, displayName, nameFull, accountId, cop
                 Platform.current.copyString(copyName);
             }}
         >
-            <Label basic image className="user-label">
-                {avatar && !!displayName && <Image className="av-class" avatar spaced="right" src={avatar} />}
+            <span className={styles.label}>
+                {avatar && !!displayName && <img className={styles.avatar} src={avatar} alt="" />}
                 <HighlightenText text={displayName} />
-            </Label>
+            </span>
         </span>
     );
 }
