+++ src/components/pull-requests/PullRequestRow.tsx
@@ -7,9 +7,9 @@ import { PullRequest } from "../../models/pull-request";
 import { PullRequestReviewer } from "../../models/pull-request-reviewer";
 import { s } from "../../values/Strings";
 import { Link } from "../Link";
-import { ProfileWidget } from "../ProfileWidget";
+import { ProfileWidget } from "../profile-widget/profile-widget";
 import { Tag } from "../Tag";
-import { PRReviewer } from "./PRReviewer";
+import { PRReviewer } from "./pr-reviewer/pr-reviewer";
 import { PullRequestContextMenu } from "./PullRequestContextMenu";
 import { getApi } from "/@/api/client";
 import { useDataStore } from "/@/zustand/data";
