+++ src/components/pull-requests/pr-reviewer/pr-reviewer.module.css
@@ -1,0 +1,49 @@
+.label {
+    display: inline-flex;
+    align-items: center;
+    cursor: default;
+    border: none;
+    background: transparent;
+    font-weight: normal;
+    box-shadow: none;
+    position: relative;
+    padding: 0;
+    vertical-align: bottom;
+}
+
+.avatarWrap {
+    display: inline-block;
+    position: relative;
+    width: 26px;
+    height: 26px;
+    border-radius: 50%;
+    margin-right: 4px;
+    flex-shrink: 0;
+    overflow: hidden;
+}
+
+.avatar {
+    width: 100%;
+    height: 100%;
+    border-radius: 50%;
+    object-fit: cover;
+    display: block;
+}
+
+.requiredBorder::after {
+    content: "";
+    position: absolute;
+    inset: 0;
+    border-radius: 50%;
+    border: 2px solid #21cfff;
+    pointer-events: none;
+}
+
+.statusIcon {
+    position: absolute;
+    right: -2px;
+    bottom: 0;
+    background-color: white;
+    border-radius: 50%;
+    z-index: 1;
+}
