+++ src/components/pull-requests/pr-reviewer/pr-reviewer.tsx
@@ -1,9 +1,8 @@
 import { Icon } from "semantic-ui-react";
-import { PullRequestReviewer } from "../../models/pull-request-reviewer";
-import { Label } from "../../ui/label";
+import { PullRequestReviewer } from "../../../models/pull-request-reviewer";
 import { useAvatar } from "/@/hooks/useAvatar";
-import { Image } from "/@/ui/image";
 import { s } from "/@/values/Strings";
+import styles from "./pr-reviewer.module.css";
 
 interface P {
     accountId: string;
@@ -17,21 +16,21 @@ export function PRReviewer({ reviewer, accountId }: P) {
         const color = vote > 0 ? "green" : vote < -5 ? "red" : "orange";
         const iconName = vote > 0 ? "check circle" : vote < -5 ? "times circle" : "clock";
 
-        return <Icon className="pr-reviewer-status" name={iconName} color={color} />;
+        return <Icon className={styles.statusIcon} name={iconName} color={color} />;
     };
 
     const avatar = useAvatar(accountId, reviewer.imageUrl);
 
     return (
         <span title={reviewer.name + (reviewer.isRequired ? ` (${s("requiredReviewer")})` : "")}>
-            <Label basic image className="user-label">
-                <Image
-                    className={`av-class${reviewer.isRequired ? " pr-reviewer-required" : ""}`}
-                    avatar
-                    src={avatar}
-                />
+            <span className={styles.label}>
+                <span
+                    className={`${styles.avatarWrap}${reviewer.isRequired ? ` ${styles.requiredBorder}` : ""}`}
+                >
+                    <img className={styles.avatar} src={avatar ?? undefined} alt="" />
+                </span>
                 {getStatusIcon(reviewer.vote)}
-            </Label>
+            </span>
         </span>
     );
 }
